test(workflows): move coerceToSuffix tests to data_suffix_test.ts
Per CLAUDE.md, tests live next to their source — `foo.ts → foo_test.ts`.
After Commit 5 of #1252 extracted coerceToSuffix into data_suffix.ts,
the 9 tests covering it remained in execution_service_test.ts and were
labeled "coerceToSuffix tests". Move them next to the source they
exercise.

No new tests; no test deleted. Same 9 cases, identical expectations.
Total suite count unchanged: 5045 passed.
